==> notebooks/2_Data_Preprocessing.ipynb <==
# 2 - Data Preprocessing


## Import Libraries

import pandas as pd

## Load Datasets


# Read in the dataframes
hot100_df = pd.read_csv('../data/1_hot100_raw.csv', sep=';')
mss_df = pd.read_csv('../data/2_million_song_subset_raw.csv', sep=';')

# Show the first 5 rows and shape of each dataframe
display(hot100_df.head())
print(hot100_df.shape)

display(mss_df.head())
print(mss_df.shape)

## Clean MillionSongSubset df

# Create a copy of the dataframe
mss_clean_df = mss_df.copy()

# Define column type as string
mss_clean_df['title'] = mss_clean_df['title'].astype(str)
mss_clean_df['artist'] = mss_clean_df['artist'].astype(str)

# Clean titles and artists by replacing extra characters: "'" and "b"
mss_clean_df['title'] = mss_clean_df['title'].str.replace('b|\'', '', regex = True)
mss_clean_df['artist'] = mss_clean_df['artist'].str.replace('b|\'', '', regex = True)

# Show first 5 rows
mss_clean_df.head()

## Combine Datasets & Drop Duplicate Rows

# Create a new column called 'is_hot' for both dataframes
mss_clean_df['is_hot'] = False
hot100_df['is_hot'] = True

# Concatenate the two dataframes, ignoring the index
combined_df = pd.concat([mss_clean_df, hot100_df], ignore_index=True)

# Drop duplicates
combined_df = combined_df.drop_duplicates(subset=['artist', 'title'])

# Show shape
combined_df.shape

## Save Combined Clean Dataset

# Save the dataframe to a csv file
combined_df.to_csv('../data/3_combined_clean_dataset.csv', sep=';', index=False)
