==> notebooks/3_Data_Extension_SpotifyAPI.ipynb <==
# Data Extension Via Spotify API

In this notebook, we will extend the features of the combined dataset using the Spotify API. This will be done in 2 steps:
1. Develop Spotify data fetch functions

2. Run functions on the combined, clean dataset

## Import Libraries

import os
import random
import time
import pprint
import pandas as pd
import spotipy
import json

from IPython.display import clear_output
from spotipy.oauth2 import SpotifyClientCredentials

# Define the root path
root_path = '../'

# Change the current working directory to the root path
os.chdir(root_path)

from credentials import *
from utils import functions

## Load Data

# Load the dataset
df = pd.read_csv('data/3_combined_clean_dataset.csv', sep=';')

# Create a sample of 20 songs for testing
test_df = df.sample(n=20, random_state=42)

## 1. Spotify Data Fetch & Utility Functions

### Function search_song()

# Create a function to search a given single song in the Spotify API: search_song(title, artist).

def search_song(title: str, artist: str) -> list:
    '''
    Given a song title and artist, this function returns up to 5 posible matches
    from Spotipy, Spotify's API.
    
    Input:
    title: song title as string
    artist: song artist as string
    
    Output:
    List of matches, each match is a dictionary
    
    '''
    
    # Initialize SpotiPy with user credentias
    sp = spotipy.Spotify(auth_manager=SpotifyClientCredentials(client_id=Client_ID,
                                                           client_secret=Client_Secret))

    # Create the query string
    query = ("track:"+title if title != "" else "")+" "+("artist:"+artist if artist != "" else "")

    # Search the requested song and artist if the input is not blank
    if len(query) > 0:
        #try:
            raw_results = sp.search(q=query,type="track,artist",limit=5)
            raw_results = raw_results['tracks']['items'] # Parse first two dictionary levels
        #except Exception as e:
            #print("Error:", e)
    else:
        raw_results = []
    
    # Define result variable
    clean_results = []
    
    # Loop through results and store relevant data in clean_results as list of dictionaries
    for index, result in enumerate(raw_results):
        
        # Store each individual result in temporal dictionary             TO-DO: add -> song poopularity, artist nationality...
        clean_result = {'result_index': index,
                        'track_id': result['id'],
                        'href': result['href'],
                        'track_link': result['external_urls']['spotify'],
                        'track_name': result['name'],
                        'artist': result['album']['artists'][0]['name'],
                        'album_name': result['album']['name'],
                        'album_release_year': result['album']['release_date'],
                        'track_duration': round(result['duration_ms']/1000),
                        'album_image': result['album']['images'][1]['url'],
                        'popularity': result['popularity'],
        }
        
        # Append each dictionary to results list
        clean_results.append(clean_result)
    
    return clean_results

# Test function
search_song('Scar Tissue', 'Red Hot')

### Function bulk_song_search()

# Create a function to get all the Spotify song data

def bulk_song_search(df: pd.DataFrame) -> pd.DataFrame:
    '''
    Given a DataFrame with two columns: title and artist,
    this function returns a new df with track data 
    using Spotipy, Spotify's API.
    
    Input:
    df: pandas DataFrame with songs
    
    Output:
    DataFrame with the track data
    
    '''
    
    # Define loop pauses to avoid API server blockage by overflow
    sleep_time = 30 # seconds of sleep time
    loop_count = 50 # tracks per sleep loop
    
    # Define an empty array to append results
    sp_songs = []
    
    loop_counter = 0 # counter to track loop
    counter = 0 # counter to track progress
    fail_counter = 0 # counter to track failed matches
                 
    for index, row in df.iterrows():
        
        counter += 1 # Add one to the global song counter
        
        # Clear last print and print next progress
        clear_output(wait=True)
        print('Downloading song Data...', round(((counter + 1) / df.shape[0])*100), '%')
        
        # Get the current row song title and artist. Fix to maximum 30 chars to avoid API error.
        song_name = str(row['title'])[:30]
        artist_name = str(row['artist'])[:30]
        
        try:
            sp_data = search_song(song_name, artist_name) # Call search_song function to get data
            if len(sp_data) > 0:
                sp_data[0]['is_hot'] = row['is_hot'] # Add is_hot key-value
                sp_songs.append(sp_data[0]) # If there is data, append to global variable
            else:
                fail_counter += 1 # If there is no data, add 1 to fail_counter
        except Exception as e:
            # Log error and add counter
            print("Error:", e)
            fail_counter += 1
        
        # Sleep to avoid API overflow...
        loop_counter += 1
        if loop_counter >= loop_count:
            loop_counter = 0
            print('Sleeping for', sleep_time, 's...') # Print sleep time
            time.sleep(sleep_time) # Sleep -> pause download to simulate human behavior

    # Print download summary
    clear_output(wait=True)
    print("Song data download complete.", "Success rate:", round(((counter-fail_counter)/counter)*100), "%")
    print("Succesful downloads:",counter-fail_counter,"Failed downloads:",fail_counter)
    
    # Convert list of dictionaries to DataFrame
    sp_songs_df = pd.DataFrame(sp_songs)
    
    # Drop result_index column: used for single songs searches in search_song() function
    sp_songs_df.drop(inplace=True, columns='result_index')
    
    return sp_songs_df

# Test function
bulk_song_search_test_df = bulk_song_search(test_df)

### Function get_audio_features()

+ [Description of Audio Features](https://developer.spotify.com/documentation/web-api/reference/#/operations/get-several-audio-features)

# Create a function to search a given single song in the Spotify API: search_song(title, artist).

def get_audio_features(track_ids: list) -> pd.DataFrame:
    '''
    Given a list of track IDs, this function returns 
    track audio features such as danceability, energy or tempo
    using Spotipy, Spotify's API.
    
    Input:
    track_ids: list of Spotipy track_ids
    
    Output:
    DataFrame with the track_ids and audio features
    
    '''
    
    # Define loop pauses to avoid API server blockage by overflow
    sleep_time = 30 # seconds of sleep time
    loop_count = 50 # tracks per sleep loop
    
    # Initialize SpotiPy with user credentias
    sp = spotipy.Spotify(auth_manager=SpotifyClientCredentials(client_id=Client_ID,
                                                           client_secret=Client_Secret))

    # Define result variable
    clean_results = []
    
    loop_counter = 0 # counter to track loop
    counter = 0 # counter to track progress
    fail_counter = 0 # counter to track failed matches
    
    # Loop through track_ids to get relevant audio features
    for index, track_id in enumerate(track_ids):
        
        counter += 1 # Add 1 to global counter
        clear_output(wait=True) # Clear print output to update progress
        print('Downloading audio features...', round(((index + 1) / len(track_ids))*100), '%') # Print progress
        
        try:
            # Get audio features from Spotipy
            api_result = sp.audio_features(track_id)[0]

            # Add features to track variable
            track = {'track_id': track_id,
                     'acousticness': api_result['acousticness'],
                     'danceability': api_result['danceability'],
                     'energy': api_result['energy'],
                     'instrumentalness': api_result['instrumentalness'],
                     'key': api_result['key'],
                     'liveness': api_result['liveness'],
                     'loudness': api_result['loudness'],
                     'mode': api_result['mode'],
                     'speechiness': api_result['speechiness'],
                     'tempo': api_result['tempo'],
                     'time_signature': api_result['time_signature'],
                     'valence': api_result['valence'],
            }

            # Append track results to results list
            clean_results.append(track)

        except Exception as e:
            # If there is an error:
            print("Error:", e)
            fail_counter += 1
        
        # Sleep to avoid API overflow...
        loop_counter += 1
        if loop_counter >= loop_count:
            loop_counter = 0
            print('Sleeping for', sleep_time, 's...')
            time.sleep(sleep_time) # Sleep
    
    clear_output(wait=True)
    print("Audio features download complete.", "Success rate:", round(((counter-fail_counter)/counter)*100), "%")
    print("Succesful downloads:",counter-fail_counter,"Failed downloads:",fail_counter)
    
    # Convert list of dictionaries to DataFrame
    clean_results_df = pd.DataFrame(clean_results)

    return clean_results_df



# Create list of track_ids from the Spotify song dataset
track_ids = list(bulk_song_search_test_df['track_id'])

# Test function
audio_features_test_df = get_audio_features(track_ids)

### Function add_audio_features()

# Create function to concatenate dataframes

def add_audio_features(df1: pd.DataFrame, df2: pd.DataFrame) -> pd.DataFrame:
    '''
    This function concatenates the audio features columns to the base 
    song DataFrame columns and removes any duplicate columns of the 
    concatenates DataFrame.
    
    Input:
    df: base songs DataFrame
    audio_features
    
    '''
    
    # Merge datasets based on track_id column
    merged_df = pd.merge(df1, df2, on='track_id', how='inner')
    
    # Remove duplicate columns
    merged_df = merged_df.loc[:, ~merged_df.columns.duplicated()]
    
    return merged_df



# Test function
concat_df = add_audio_features(bulk_song_search_test_df, audio_features_test_df)

# Show concat dataset
concat_df.head()

## 2. Run Functions & Save Data

# Download song data
sp_songs_df = bulk_song_search(df)

# Remove rows with duplicate track_id
sp_songs_df = sp_songs_df[~sp_songs_df.duplicated(subset='track_id', keep='first')]

# Save data as csv
sp_songs_df.to_csv('./data/4_sp_songs.csv',index = False, sep=";")

# Create list of track_ids from all song data
track_ids = list(sp_songs_df['track_id'])

# Download audio features data
audio_features_df = get_audio_features(track_ids)

# Save data as csv
audio_features_df.to_csv('./data/5_sp_audio_features.csv',index = False, sep=";")

# Merge song + audio features data
merge_df = add_audio_features(sp_songs_df, audio_features_df)

# Save data as csv
merge_df.to_csv('./data/6_sp_combined_dataset.csv',index = False, sep=";")

# Show dfs shape
print(sp_songs_df.shape)
print(audio_features_df.shape)
print(merge_df.shape)

# Show concat dataset
merge_df.head()
