==> notebooks/5_Song_Recommender.ipynb <==
# Song Recommender Project

## Import libraries

import os
import pandas as pd

# Define the root path
root_path = '../'

# Change the current working directory to the root path
os.chdir(root_path)

import utils.functions as f

## Model selection

# Select model
selected_model_name = 'kmeans_17' # 'kmeans_8'

## Define user input settings

# Title & artist input by user
user_input_title = 'Last Night'
user_input_artist = ''

## Get input song data

# Get spotify data including track_id
sp_song_data = f.search_song(user_input_title, user_input_artist)

## User song selection

# User selects index of Spotify search results
user_song_selection = 0
track_id = sp_song_data[user_song_selection]['track_id']

## Song recommendation

recommended_songs = f.recommend_song(selected_model_name, track_id, 0, 100)

recommended_songs
